Create fisher_location.r
This is a simple script for R that prompts the user for rates of conformity to a prescription protocol by clinic location, and then performs a Fisher test on that data. It will be a useful tool to learn some for advanced techniques. I want to be able to:

i) Implement a loop so that once the script stores the statistics it will return to the top for the next batch, rather than having to be run again.
ii) Store the data for each batch in unique variables, rather than overwriting the one variable each time the loop iterates
iii) Writes the data to a file in an appropriate (i.e., legible) format